+++ compatibility_server/configs.py
@@ -23,43 +23,6 @@ in this file since they exist in the same workflow.
 """
 
 
-import enum
-
-
-class PackageStatus(enum.Enum):
-    """Represents a package's compatibility status.
-
-    The status is based on the results of running 'pip install' and
-    'pip check' on the compatibility server
-
-    UNKNOWN_PACKAGE: package not in whitelist
-    INTERNAL_ERROR: unexpected internal error
-    MISSING_DATA: missing package data from package store
-    SELF_INCOMPATIBLE: pip error when installing self
-    INCOMPATIBLE: pip error when installed with another package
-    OUTDATED_DEPENDENCY: package has an outdated dependency
-    SUCCESS: No issues
-    """
-    UNKNOWN_PACKAGE = 'UNKNOWN_PACKAGE'
-    INTERNAL_ERROR = 'INTERNAL_ERROR'
-    MISSING_DATA = 'MISSING_DATA'
-    SELF_INCOMPATIBLE = 'SELF_INCOMPATIBLE'
-    INCOMPATIBLE = 'INCOMPATIBLE'
-    OUTDATED_DEPENDENCY = 'OUTDATED_DEPENDENCY'
-    SUCCESS = 'SUCCESS'
-
-
-PACKAGE_STATUS_TO_COLOR = {
-    PackageStatus.UNKNOWN_PACKAGE: 'lightgrey',
-    PackageStatus.INTERNAL_ERROR: 'lightgrey',
-    PackageStatus.MISSING_DATA: 'lightgrey',
-    PackageStatus.SELF_INCOMPATIBLE: 'red',
-    PackageStatus.INCOMPATIBLE: 'red',
-    PackageStatus.OUTDATED_DEPENDENCY: 'orange',
-    PackageStatus.SUCCESS: 'green',
-}
-
-
 def _format_url(repo_name, setuppy_path=''):
     url = 'git+git://github.com/{}.git'.format(repo_name)
     if setuppy_path != '':
